Docs: finish features overview page.
